Extract the NWConnection code into StateManagedNWConnectionChannel
